4. Laboratorijska vježba

+++ api/or_api.py
@@ -48,6 +48,9 @@ def apply_caching(response):
 def get_all_phones():
     cursor = db.get_cursor()
 
+    if cursor == False:
+        return util.build_response({"error": "Service is currently unavailable"}), HTTPStatus.SERVICE_UNAVAILABLE
+
     cursor.execute("SELECT * FROM phone;")
     data = cursor.fetchall()
     cursor.close()
@@ -59,6 +62,9 @@ def get_all_phones():
 def get_phone(phone_id):
     cursor = db.get_cursor()
 
+    if cursor == False:
+        return util.build_response({"error": "Service is currently unavailable"}), HTTPStatus.SERVICE_UNAVAILABLE
+
     cursor.execute("SELECT * FROM phone WHERE id = %s;", (phone_id,))
     data = cursor.fetchall()
     cursor.close()
@@ -66,10 +72,15 @@ def get_phone(phone_id):
     if data is None:
         return util.build_response({"error": "Invalid request"}), HTTPStatus.BAD_REQUEST
 
-    if not data:
-        return util.build_response({"success": True, "response": data}), HTTPStatus.NOT_FOUND
+    context = {
+        "name": "https://schema.org/name",
+        "release_date": "https://schema.org/Date"
+    }
 
-    return util.build_response({"success": True, "response": data}), HTTPStatus.OK
+    if not data:
+        return util.build_response({"success": True, "response": data, "context": context}), HTTPStatus.NOT_FOUND
+
+    return util.build_response({"success": True, "response": data, "context": context}), HTTPStatus.OK
 
 # -------------------------------- Companies --------------------------------
 
@@ -77,6 +88,9 @@ def get_phone(phone_id):
 @app.route('/company', methods=['GET'])
 def get_all_companies():
     cursor = db.get_cursor()
+
+    if cursor == False:
+        return util.build_response({"error": "Service is currently unavailable"}), HTTPStatus.SERVICE_UNAVAILABLE
 
     cursor.execute("SELECT * FROM company;")
     data = cursor.fetchall()
@@ -89,6 +103,9 @@ def get_all_companies():
 def get_company(company_id):
     cursor = db.get_cursor()
 
+    if cursor == False:
+        return util.build_response({"error": "Service is currently unavailable"}), HTTPStatus.SERVICE_UNAVAILABLE
+
     cursor.execute("SELECT * FROM company WHERE id = %s;", (company_id,))
     data = cursor.fetchall()
     cursor.close()
@@ -96,10 +113,14 @@ def get_company(company_id):
     if data is None:
         return util.build_response({"error": "Invalid request"}), HTTPStatus.BAD_REQUEST
 
-    if not data:
-        return util.build_response({"success": True, "response": data}), HTTPStatus.NOT_FOUND
+    context = {
+        "name": "https://schema.org/name"
+    }
 
-    return util.build_response({"success": True, "response": data}), HTTPStatus.OK
+    if not data:
+        return util.build_response({"success": True, "response": data, "context": context}), HTTPStatus.NOT_FOUND
+
+    return util.build_response({"success": True, "response": data, "context": context}), HTTPStatus.OK
 
 
 @app.route('/company', methods=['POST'])
@@ -115,6 +136,9 @@ def create_comapny():
         return util.build_response({"error": "Field 'name' must not be empty"}), HTTPStatus.BAD_REQUEST
 
     cursor = db.get_cursor()
+
+    if cursor == False:
+        return util.build_response({"error": "Service is currently unavailable"}), HTTPStatus.SERVICE_UNAVAILABLE
 
     company_name = request.json['name']
 
@@ -144,6 +168,9 @@ def update_company(company_id):
 
     cursor = db.get_cursor()
 
+    if cursor == False:
+        return util.build_response({"error": "Service is currently unavailable"}), HTTPStatus.SERVICE_UNAVAILABLE
+
     company_name = request.json['name']
 
     cursor.execute("UPDATE company SET name=%s WHERE id=%s",
@@ -156,6 +183,9 @@ def update_company(company_id):
 @app.route('/company/<int:company_id>', methods=['DELETE'])
 def delete_company(company_id):
     cursor = db.get_cursor()
+
+    if cursor == False:
+        return util.build_response({"error": "Service is currently unavailable"}), HTTPStatus.SERVICE_UNAVAILABLE
     cursor.execute("DELETE FROM company WHERE id = %s", (company_id, ))
     db.commit()
 
@@ -169,6 +199,9 @@ def delete_company(company_id):
 def get_all_processors():
     cursor = db.get_cursor()
 
+    if cursor == False:
+        return util.build_response({"error": "Service is currently unavailable"}), HTTPStatus.SERVICE_UNAVAILABLE
+
     cursor.execute("SELECT * FROM processor;")
     data = cursor.fetchall()
     cursor.close()
@@ -180,6 +213,9 @@ def get_all_processors():
 def get_processor(processor_id):
     cursor = db.get_cursor()
 
+    if cursor == False:
+        return util.build_response({"error": "Service is currently unavailable"}), HTTPStatus.SERVICE_UNAVAILABLE
+
     cursor.execute("SELECT * FROM processor WHERE id = %s;", (processor_id,))
     data = cursor.fetchall()
     cursor.close()
@@ -187,10 +223,15 @@ def get_processor(processor_id):
     if data is None:
         return util.build_response({"error": "Invalid request"}), HTTPStatus.BAD_REQUEST
 
-    if not data:
-        return util.build_response({"success": True, "response": data}), HTTPStatus.NOT_FOUND
+    context = {
+        "brand": "https://schema.org/brand",
+        "name": "https://schema.org/name"
+    }
 
-    return util.build_response({"success": True, "response": data}), HTTPStatus.OK
+    if not data:
+        return util.build_response({"success": True, "response": data, "context": context}), HTTPStatus.NOT_FOUND
+
+    return util.build_response({"success": True, "response": data, "context": context}), HTTPStatus.OK
 
 # -------------------------------- Cameras --------------------------------
 
@@ -198,6 +239,9 @@ def get_processor(processor_id):
 @app.route('/camera', methods=['GET'])
 def get_all_cameras():
     cursor = db.get_cursor()
+
+    if cursor == False:
+        return util.build_response({"error": "Service is currently unavailable"}), HTTPStatus.SERVICE_UNAVAILABLE
 
     cursor.execute("SELECT * FROM camera;")
     data = cursor.fetchall()
@@ -210,6 +254,9 @@ def get_all_cameras():
 def get_camera(camera_id):
     cursor = db.get_cursor()
 
+    if cursor == False:
+        return util.build_response({"error": "Service is currently unavailable"}), HTTPStatus.SERVICE_UNAVAILABLE
+
     cursor.execute("SELECT * FROM camera WHERE id = %s;", (camera_id,))
     data = cursor.fetchall()
     cursor.close()
@@ -217,10 +264,16 @@ def get_camera(camera_id):
     if data is None:
         return util.build_response({"error": "Invalid request"}), HTTPStatus.BAD_REQUEST
 
-    if not data:
-        return util.build_response({"success": True, "response": data}), HTTPStatus.NOT_FOUND
+    context = {
+        "description": "https://schema.org/description",
+        "vertical_resolution": "https://schema.org/Integer",
+        "horizontal_resolution": "https://schema.org/Integer"
+    }
 
-    return util.build_response({"success": True, "response": data}), HTTPStatus.OK
+    if not data:
+        return util.build_response({"success": True, "response": data, "context": context}), HTTPStatus.NOT_FOUND
+
+    return util.build_response({"success": True, "response": data, "context": context}), HTTPStatus.OK
 
 
 @app.route('/documentation', methods=['GET'])
